paper-translation/requirements.txt
pdfplumber>=0.10.0
pymupdf>=1.23.0
reportlab>=4.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
